add packages for every class
