+++ app/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     compile 'com.google.firebase:firebase-core:15.0.0'
     compile 'com.google.firebase:firebase-auth:15.0.0'
     compile 'com.google.firebase:firebase-database:15.0.0'
+    compile 'com.google.firebase:firebase-storage:15.0.0'
     compile 'com.firebaseui:firebase-ui:3.3.1'
 
             compile 'com.google.android.gms:play-services-auth:15.0.0'
