+++ app/src/main/AndroidManifest.xml
@@ -11,15 +11,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SignUp" />
-        <activity android:name=".LogIn"></activity>
-        <activity android:name=".AdminDashboardActivity" />
+        <activity android:name=".activities.SignUp" />
+        <activity android:name=".activities.LogIn"></activity>
+        <activity android:name=".activities.AdminDashboardActivity" />
 
         <meta-data android:name="com.facebook.sdk.ApplicationId" android:value="@string/facebook_app_id" tools:replace="android:value"/>
         <activity android:name="com.facebook.FacebookActivity"
