+++ app/src/main/java/com/shuzhongchen/foodordersystem/activities/LogIn.java
@@ -1,21 +1,20 @@
-package com.shuzhongchen.foodordersystem;
+package com.shuzhongchen.foodordersystem.activities;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Switch;
-import android.widget.Toast;
 
 import com.facebook.CallbackManager;
 import com.facebook.FacebookCallback;
 import com.facebook.FacebookException;
 import com.facebook.login.LoginResult;
 import com.facebook.login.widget.LoginButton;
+import com.shuzhongchen.foodordersystem.R;
+import com.shuzhongchen.foodordersystem.activities.AdminDashboardActivity;
 
 import java.util.Arrays;
