+++ app/src/main/java/com/shuzhongchen/foodordersystem/activities/MainActivity.java
@@ -1,18 +1,12 @@
-package com.shuzhongchen.foodordersystem;
+package com.shuzhongchen.foodordersystem.activities;
 
 import android.content.Intent;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.Signature;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Base64;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
+import com.shuzhongchen.foodordersystem.R;
 
 public class MainActivity extends AppCompatActivity {
