+++ app/src/main/java/com/shuzhongchen/foodordersystem/holders/CustomListView.java
@@ -1,16 +1,20 @@
-package com.shuzhongchen.foodordersystem;
+package com.shuzhongchen.foodordersystem.holders;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
+import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
-import java.util.ArrayList;
+import com.shuzhongchen.foodordersystem.R;
+
+import info.hoang8f.widget.FButton;
 
 /**
  * Created by jinchengcheng on 4/29/18.
@@ -80,6 +84,46 @@ public class CustomListView extends ArrayAdapter<String>{
             UnitPriceTV = (TextView) view.findViewById(R.id.menu_unit_price_textView);
             PrepTimeTV = (TextView) view.findViewById(R.id.menu_preptime_textView);
             imageButton = (ImageButton) view.findViewById(R.id.menu_image);
+        }
+    }
+
+    /**
+     * Created by jinchengcheng on 5/1/18.
+     */
+
+    public static class MenuViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
+
+        public TextView nameTV;
+        public TextView categoryTV;
+        public TextView caloriesTV;
+        public TextView UnitPriceTV;
+        public TextView PrepTimeTV;
+
+        public ImageButton imageButton;
+        public Button editButton;
+        public Button removeButton;
+
+
+
+        //private ItemClickListener itemClickListener;
+
+        public MenuViewHolder(View view) {
+            super(view);
+
+            nameTV = (TextView)view.findViewById(R.id.menu_name_textView);
+            categoryTV = (TextView)view.findViewById(R.id.menu_category_textview);
+            caloriesTV = (TextView) view.findViewById(R.id.menu_calories_textView);
+            UnitPriceTV = (TextView) view.findViewById(R.id.menu_unit_price_textView);
+            PrepTimeTV = (TextView) view.findViewById(R.id.menu_preptime_textView);
+            imageButton = (ImageButton) view.findViewById(R.id.menu_image);
+
+            editButton = (FButton)view.findViewById(R.id.edit_fbutton);
+            removeButton = (FButton) view.findViewById(R.id.remove_fbutton);
+        }
+
+        @Override
+        public void onClick(View view) {
+
         }
     }
 }
