+++ app/src/main/res/layout/activity_log_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/login_background"
-    tools:context="com.shuzhongchen.foodordersystem.LogIn">
+    tools:context="com.shuzhongchen.foodordersystem.activities.LogIn">
 
     <LinearLayout
         android:id="@+id/linearLayout"
