+++ app/src/main/res/layout/activity_menu_detail.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/login_background"
     android:orientation="vertical"
-    tools:context="com.shuzhongchen.foodordersystem.Admin">
+    tools:context="com.shuzhongchen.foodordersystem.models.Admin">
 
 
     <LinearLayout
