+++ app/src/main/res/layout/activity_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/signup_backgrand"
-    tools:context="com.shuzhongchen.foodordersystem.SignUp">
+    tools:context="com.shuzhongchen.foodordersystem.activities.SignUp">
 
     <LinearLayout
         android:layout_width="match_parent"
